requirements.txt
aiogram==2.25.1
requests
python-dotenv
yt-dlp
spotipy
